Separate shutdown test; it leaves Stairway down. Various other minor fixes
