+++ src/test/java/bio/terra/service/job/JobServiceShutdownTest.java
@@ -1,0 +1,83 @@
+package bio.terra.service.job;
+
+import bio.terra.app.configuration.ApplicationConfiguration;
+import bio.terra.common.category.Unit;
+import bio.terra.service.iam.AuthenticatedUserRequest;
+import bio.terra.stairway.FlightState;
+import bio.terra.stairway.FlightStatus;
+import bio.terra.stairway.Stairway;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+
+import static org.junit.Assert.assertTrue;
+
+@RunWith(SpringRunner.class)
+@AutoConfigureMockMvc
+@SpringBootTest
+@Category(Unit.class)
+public class JobServiceShutdownTest {
+    private static final Logger logger = LoggerFactory.getLogger(JobServiceShutdownTest.class);
+
+    private AuthenticatedUserRequest testUser = new AuthenticatedUserRequest()
+        .subjectId("StairwayUnit")
+        .email("stairway@unit.com")
+        .token(Optional.empty());
+
+    @Autowired
+    private JobService jobService;
+
+    @Autowired
+    private ApplicationConfiguration appConfig;
+
+    @Test
+    public void testShutdown() throws Exception {
+        // Segregated into its own test, since the Stairway object is not usable after it completes
+        // scenario:
+        //  - start a thread that creates one flight every X second; keeping track of the flights
+        //     - each flight sleeps Y secs and then returns; we want to get a backlog of flights in the queue
+        //    when thread catches JobServiceShutdown exception, it stops launching and returns its flight list
+        //  - sleep Z seconds to let the system fill up
+        //  - call shutdown
+        //  - validate that flights are either SUCCESS or READY
+
+        // Test Control Parameters
+        final int flightPerSecond = 1;
+        final int flightWaitSeconds = 30;
+
+        int maxStairwayThreads = appConfig.getMaxStairwayThreads();
+        logger.info("maxStairwayThreads: " + maxStairwayThreads);
+
+        JobTestShutdownFlightLauncher launcher =
+            new JobTestShutdownFlightLauncher(flightPerSecond, flightWaitSeconds, testUser, jobService);
+
+        Thread launchThread = new Thread(launcher);
+        launchThread.start();
+
+        // Build a backlog in the queue
+        TimeUnit.SECONDS.sleep(2 * maxStairwayThreads);
+        jobService.shutdown();
+
+        launchThread.join();
+        List<String> flightIdList = launcher.getFlightIdList();
+        Stairway stairway = jobService.getStairway();
+
+        for (String flightId : flightIdList) {
+            FlightState state = stairway.getFlightState(flightId);
+            FlightStatus status = state.getFlightStatus();
+            logger.info("Flightid: " + flightId + "; status: " + status);
+            assertTrue(status == FlightStatus.SUCCESS || status == FlightStatus.READY);
+        }
+    }
+
+}
