+++ src/test/java/bio/terra/service/job/JobServiceTest.java
@@ -4,13 +4,9 @@ import bio.terra.app.configuration.ApplicationConfiguration;
 import bio.terra.common.category.Unit;
 import bio.terra.model.JobModel;
 import bio.terra.service.iam.AuthenticatedUserRequest;
-import bio.terra.stairway.FlightState;
-import bio.terra.stairway.FlightStatus;
-import bio.terra.stairway.Stairway;
 import bio.terra.stairway.exception.FlightNotFoundException;
 import org.broadinstitute.dsde.workbench.client.sam.model.ResourceAndAccessPolicy;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
@@ -25,11 +21,9 @@ import org.springframework.test.context.junit4.SpringRunner;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.TimeUnit;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringRunner.class)
 @AutoConfigureMockMvc
@@ -48,11 +42,6 @@ public class JobServiceTest {
 
     @Autowired
     private ApplicationConfiguration appConfig;
-
-    @Before
-    public void setup() throws Exception {
-
-    }
 
     @Test
     public void retrieveTest() throws Exception {
@@ -135,45 +124,6 @@ public class JobServiceTest {
     @Test(expected = FlightNotFoundException.class)
     public void testBadIdRetrieveResult() throws InterruptedException {
         jobService.retrieveJobResult("abcdef", Object.class, null);
-    }
-
-    @Test
-    public void testShutdown() throws Exception {
-        // scenario:
-        //  - start a thread that creates one flight every X second; keeping track of the flights
-        //     - each flight sleeps Y secs and then returns; we want to get a backlog of flights in the queue
-        //    when thread catches JobServiceShutdown exception, it stops launching and returns its flight list
-        //  - sleep Z seconds to let the system fill up
-        //  - call shutdown
-        //  - validate that flights are either SUCCESS or READY
-
-        // Test Control Parameters
-        final int flightPerSecond = 1;
-        final int flightWaitSeconds = 30;
-
-        int maxStairwayThreads = appConfig.getMaxStairwayThreads();
-        logger.info("maxStairwayThreads: " + maxStairwayThreads);
-
-        JobTestShutdownFlightLauncher launcher =
-            new JobTestShutdownFlightLauncher(flightPerSecond, flightWaitSeconds, testUser, jobService);
-
-        Thread launchThread = new Thread(launcher);
-        launchThread.start();
-
-        // Build a backlog in the queue
-        TimeUnit.SECONDS.sleep(2 * maxStairwayThreads);
-        jobService.shutdown();
-
-        launchThread.join();
-        List<String> flightIdList = launcher.getFlightIdList();
-        Stairway stairway = jobService.getStairway();
-
-        for (String flightId : flightIdList) {
-            FlightState state = stairway.getFlightState(flightId);
-            FlightStatus status = state.getFlightStatus();
-            logger.info("Flightid: " + flightId + "; status: " + status);
-            assertTrue(status == FlightStatus.SUCCESS || status == FlightStatus.READY);
-        }
     }
 
     private void validateJobModel(JobModel jm, int index, List<String> fids) {
